Refactor code to enhance OOP principles: Modularize classes, improve encapsulation, and separate concerns
- Split the `Max` class into multiple classes (`Ui`, `Storage`, `Parser`, `TaskList`) to follow the Single Responsibility Principle.
- Improved encapsulation by making class attributes private and providing getter/setter methods where necessary.
- Moved task-related methods to the `TaskList` class to centralize task management.
- Updated the `Storage` class to handle file I/O operations independently.
- Simplified `Parser` class logic for handling user inputs and parsing dates.
- Improved overall code readability and maintainability.
